bot_asistencia_main/requirements.txt
discord.py==2.6.4
aiomysql==0.3.2
python-dotenv==1.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pandas
openpyxl
tzdata
gspread==6.1.4
google-auth==2.36.0
cryptography
